+++ .gitignore
@@ -108,6 +108,3 @@ ENV/
 
 # mypy
 .mypy_cache/
-
-# pre-commit
---git-common-dir/
