+++ flare/env_zoo/tests/dict.txt
@@ -1,0 +1,138 @@
+$
+-
+.
+?
+!
+well
+done
+end
+time
+up
+please
+is
+your
+the
+go
+to
+navigate
+reach
+move
+could
+collect
+you
+can
+will
+destination
+target
+goal
+cat
+dog
+rooster
+bed
+sofa
+fan
+staircase
+vase
+barbecue
+bench
+cooker
+oven
+table
+wardrobe
+barrel
+bookshelf
+drawers
+screen
+trashcan
+books
+firehydrant
+plant
+umbrella
+drums
+mailbox
+tricycle
+wheelchair
+bottle
+wrong
+apples
+car
+flashlight
+photo
+sunglasses
+backpack
+carpet
+flowers
+piano
+table-tennis
+basket
+chair
+fountain
+pillow
+toilet
+basketball
+chessboard
+gift
+towel
+bathtub
+clock
+guitar
+pool-table
+train
+bike
+hair-dryer
+puzzle
+trampoline
+boiler
+headphones
+rabbit
+treadmill
+bread
+comb
+horse
+brush
+crib
+iron
+scale
+bucket
+cup
+scissors
+vacuum
+burger
+dart
+lamp
+slippers
+cake
+dog-house
+laptop
+calender
+milk
+speaker
+camera
+fence
+pan
+squeezer
+candle
+phone
+stove
+object
+near
+by
+besides
+left
+right
+front
+behind
+of
+in
+that
+do
+not
+avoid
+anything
+except
+but
+between
+and
+location
+place
+grid
